Moved ins/del from misc1 to cursor_modify.
The ins/del functions in misc1 are somewhat related to inserting and
deleting (modifying) data relative to the cursor position. By moving
these over to their own file, increased readability and maintainability
is acheived. Further refactoring is still needed to comply with nvim
coding standards.
