+++ src/nvim/cursor_modify.h
@@ -1,0 +1,9 @@
+#ifndef NVIM_CURSOR_MODIFY_H
+#define NVIM_CURSOR_MODIFY_H
+
+#include "nvim/vim.h"
+
+#ifdef INCLUDE_GENERATED_DECLARATIONS
+# include "cursor_modify.h.generated.h"
+#endif
+#endif  // NVIM_CURSOR_MODIFY_H
