+++ src/nvim/edit.c
@@ -22,6 +22,7 @@
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
 #include "nvim/cursor.h"
+#include "nvim/cursor_modify.h"
 #include "nvim/digraph.h"
 #include "nvim/eval.h"
 #include "nvim/ex_docmd.h"
